refactor: move slugify options constant into sites/brc-analytics (#1637)
